Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dlib==20.0.0
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
Flask==3.1.1
flask-cors==6.0.0
Flask-Session==0.8.0
frozenlist==1.7.0
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.4.4
numpy==2.3.0
opencv-python==4.11.0.86
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
scipy==1.15.3
sniffio==1.3.1
starlette==0.46.2
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
